+++ test/specs/methods/hide.spec.js
@@ -1,4 +1,22 @@
 describe('hide (method)', () => {
+  it('should blur focused element in viewer before setting aria-hidden', (done) => {
+    const image = window.createImage();
+    const viewer = new Viewer(image, {
+      shown() {
+        viewer.button.focus();
+        expect(viewer.viewer.ownerDocument.activeElement).to.equal(viewer.button);
+        viewer.hide(true);
+      },
+
+      hidden() {
+        expect(viewer.viewer.contains(viewer.viewer.ownerDocument.activeElement)).to.be.false;
+        done();
+      },
+    });
+
+    viewer.show();
+  });
+
   it('should hide the viewer in modal mode', (done) => {
     const image = window.createImage();
     const viewer = new Viewer(image, {